+++ NEWS.md
@@ -6,7 +6,8 @@ v0.2.1 (2016-??-??)
 
 ### New Features ###
 
-* New mathematical operations supported: `log2`, `abs2`, `fma`, `muladd`.
+* New mathematical operations supported: `log2`, `abs2`, `fma`, `muladd`,
+  `besseli`, `besselix`, `besselk`, `besselkx`.
 
 v0.2.0 (2016-07-13)
 -------------------